requirements.txt
# --- LLM / Agent ---
langchain>=1.0.0
langchain-core>=1.0.0
langchain-google-genai>=1.0.0

# --- Spotify ---
spotipy>=2.23.0

# --- Telegram Bot ---
python-telegram-bot>=20.7

# --- Environment management ---
python-dotenv>=1.0.0

# --- YouTube ---
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# --- Utilities ---
requests>=2.31.0
